% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.dichotomic.one.var.R
\name{check.dichotomic.one.var}
\alias{check.dichotomic.one.var}
\title{Verify if one specific variable of a data set is dichotomic}
\usage{
check.dichotomic.one.var(data.to.work, variable.name)
}
\arguments{
\item{data.to.work}{is a data set containing the variables to be checked.}

\item{variable.name}{is the name of a variable to be checked.}
}
\value{
TRUE or FALSE
}
\description{
This function receives a data set and the name of a specific variable and verify if it is dichotomic or not. If 'yes' then the function return TRUE.
}
\examples{
# Clean environment
closeAllConnections()
rm(list=ls())
# Set enviroment
# setwd("to your working directory")
# Load packages
library(bnpa)
# Use working data sets from package
data(dataQuantC)
head(dataQuantC)
# Show the structure of data set
str(dataQuantC)
# Set variable name
variable.name = "A"
# data set has not dichotomic variables and function will return FALSE
check.dichotomic.one.var(dataQuantC, variable.name)
# Adding dichotomic data to dataQuantC
dataQuantC$Z <- round(runif(500, min=0, max=1),0)
# Show the new structure of data set
str(dataQuantC)
# Set variable name
variable.name = "Z"
# Now data set has dichotomic variables and function will return TRUE
check.dichotomic.one.var(dataQuantC, variable.name)
}
\references{
HAYES, A F; PREACHER, K J. Statistical mediation analysis with a multicategorical independent variable. British Journal of Mathematical and Statistical Psychology, v. 67, n. 3, p. 451-470, 2014.
}
\author{
Elias Carvalho
}
