% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.package.files.R
\name{check.package.files}
\alias{check.package.files}
\title{Keep a list os packages necessary to bnpa.}
\usage{
check.package.files()
}
\value{
An error message if package not installed
}
\description{
Check if all needed packages are installed.
}
\details{
This funcion keeps a list of packages that must be installed to the correct working of bnpa and call 'check.packages" function to check if needed packages are installed.
}
\examples{
# Clean environment
closeAllConnections()
rm(list=ls())
# Load packages
library(bnpa)
# Check and install if not installed
check.package.files()
}
\author{
Elias Carvalho
}

