% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.continuous.int.to.numeric.R
\name{convert.continuous.int.to.numeric}
\alias{convert.continuous.int.to.numeric}
\title{Convert int data into numeric data.}
\usage{
convert.continuous.int.to.numeric(data.to.work)
}
\arguments{
\item{data.to.work}{is a data frame containing the variables to be converted.}
}
\value{
A new data frame with numeric type variables.
}
\description{
Convert variables of type int into numeric.
}
\details{
This function receives a dataset as parameter and convert all variables with type int 
into numeric type.
}
\examples{
# Clean environment
closeAllConnections()
rm(list=ls())
# Load packages
library(bnpa)
# Load datasets from package
data.to.work <- dataQuantC # Pre-Loaded'
head(data.to.work)
# Converting some variables to int justto test the function
data.to.work$A <- as.integer(data.to.work$A )
data.to.work$C <- as.integer(data.to.work$C )
data.to.work$G <- as.integer(data.to.work$G )
data.to.work <- convert.continuous.int.to.numeric(data.to.work)
}
\author{
Elias Carvalho
}

