% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpa.R
\docType{data}
\name{dataQualiN}
\alias{dataQualiN}
\title{A qualitative dataset to test functions}
\source{
http://www.bnlearn.com/bnrepository/
}
\usage{
data
}
\description{
This is dataset with qualitative nominal variables containing 500 obs extracted from ASIA Bayesian Networks  of bnlearn package repository.
}
\author{
Elias Carvalho, 2016-02-07
}

