% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpa.R
\docType{data}
\name{dataQuantC}
\alias{dataQuantC}
\title{A quantitative dataset to test functions}
\source{
bnlearn package
}
\usage{
data
}
\description{
This is dataset with quantitative continuous variables containing 500 obs extracted from gaussian.test from bnlearn package repository.
}
\author{
Elias Carvalho, 2016-02-07
}

