% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera.bn.pa.R
\name{gera.bn.pa}
\alias{gera.bn.pa}
\title{Learn the Bayesian Network structure from data and build a PA model}
\usage{
gera.bn.pa(
  data.to.work,
  white.list = "",
  black.list = "",
  nreplicates = 1000,
  cb.algorithms = c("gs", "iamb", "fast.iamb", "inter.iamb"),
  sb.algorithms = c("hc", "tabu"),
  outcome.var,
  build.pa = 1
)
}
\arguments{
\item{data.to.work}{is a data from which the BN structure will be learned.}

\item{white.list}{is a list of mandatory connections of BN structure to be created.}

\item{black.list}{is a list of forbiden connections of BN structure to be created.}

\item{nreplicates}{is how many times the boostrap will run.}

\item{cb.algorithms}{the name of constrained-based algorithms.}

\item{sb.algorithms}{the name of score-based algorithms.}

\item{outcome.var}{is the outcome (dependent) variable.}

\item{build.pa}{indicates if the process will bulld a PA model or not.}
}
\description{
This function receives a data set, a list of parameters to learn the BN structure based on this data set. Then with the BN ready it will build a PA model if required. The process will then save the graphs of BN and PA and PA parameters.
}
\examples{

# Clean environment
}
\references{
Scutari M (2017). Bayesian Network Constraint-Based Structure Learning Algorithms: Parallel and Optimized Implementations in the bnlearn R Package. Journal of Statistical Software, 77(2), 1-20.
}
\author{
Elias Carvalho
}
