% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.into.ordinal.R
\name{transf.into.ordinal}
\alias{transf.into.ordinal}
\title{Transform categorical variables into ordinal}
\usage{
transf.into.ordinal(data.to.work)
}
\arguments{
\item{data.to.work}{is a data set where all variables will be transformed into odered factors.}
}
\value{
The data set transformed
}
\description{
This function receives a data set with categorical variables, scan all variables and transform it into odered factors.
}
\examples{
# Clean environment
closeAllConnections()
rm(list=ls())
# Set enviroment
# setwd("to your working directory")
# Load packages
library(bnpa)
#Load Data
data(dataQualiN)
# Transform all variables into ordinal
dataQualiN <- bnpa::transf.into.ordinal(dataQualiN)
str(dataQualiN)
}
\references{
GUJARATI, Damodar N. Basic econometrics. Tata McGraw-Hill Education, 2009.
}
\author{
Elias Carvalho
}
