% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataQualiN}
\alias{dataQualiN}
\title{A qualitative data set to test functions}
\format{A data frame with 500 rows and 7 variables:
\describe{
  \item{A}{a categorical variale}
  \item{S}{a categorical variale}
  \item{T}{a categorical variale}
  \item{L}{a categorical variale}
  \item{B}{a categorical variale}
  \item{E}{a categorical variale}
  \item{X}{a categorical variale}
  \item{D}{a categorical variale}
  ...
}}
\source{
\url{http://www.bnlearn.com/bnrepository//}
}
\usage{
dataQualiN
}
\description{
This is data set with qualitative nominal variables containing 500 obs
extracted from ASIA Bayesian Networks  of bnlearn package repository.
}
\keyword{datasets}
