% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataQuantC}
\alias{dataQuantC}
\title{A quantiative data set to test functions}
\format{A data frame with 500 rows and 7 variables:
\describe{
  \item{A}{a numeric variale}
  \item{B}{a numeric variale}
  \item{C}{a numeric variale}
  \item{D}{a numeric variale}
  \item{E}{a numeric variale}
  \item{F}{a numeric variale}
  \item{G}{a numeric variale}
  ...
}}
\source{
\url{http://www.bnlearn.com/bnrepository//}
}
\usage{
dataQuantC
}
\description{
This is data set with quantiative continuous variables containing 500 obs
extracted from ASIA Bayesian Networks  of bnlearn package repository.
}
\keyword{datasets}
