% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera.bn.structure.R
\name{gera.bn.structure}
\alias{gera.bn.structure}
\title{Learn the Bayesian Network structure from data}
\usage{
gera.bn.structure(data_bn, bn.learn.algorithm, wl, bl, type.variable,
  nreplicates = 100)
}
\arguments{
\item{data_bn}{is a data frame containing the variables of the Bayesian Networks to be learned.}

\item{bn.learn.algorithm}{is a list of algorithms (present in bnlearn) to be used in Bayesian Networks learning process.}

\item{wl}{is a list with a couple of variables to build the white list.}

\item{bl}{is a list with a couple of variables to build the black list.}

\item{type.variable}{is the type of variable inside the data frame}

\item{nreplicates}{is the number os procedures to be executing during bootstraping.}
}
\value{
The Bayesian Networks strucutre learn bootstrapped
}
\description{
Receives a dataset and learn the structure of Bayesian Network (BN) diagram using algorithms implemented in bnlearn package.
}
\details{
This function receives a dataset, a list of algorithms, a list of required nodes and another list of forbidden
nodes, the type of variables in the dataset and the number of replicates the bootstrap process will execute as
parameters. Then it learn the Bayesian Network structure and return it .
}
\author{
Elias Carvalho
}

